[MA-22] fix : 매칭 상세 뷰 로직 수정

+++ domain/src/main/java/com/motax/modutaxi/domain/model/TaxiPotDetailData.kt
@@ -14,7 +14,7 @@ data class TaxiPotDetailData(
     val durationMinutes: Int,
     val expectedCharge: Int,
     val expectedChargePerPerson: Int,
-    val managerId: Int,
+    val managerId: Long,
     val myRoom: Boolean,
     val participate: Boolean,
     val path: PathData,
